Working test of find_ace_high
This is about the ugliest thing I've ever done, but this seems to work properly when it comes to getting it to run.
I'm at the very least certain of the logic for matches working properly, it's just the syntax for sending and returning Vectors that's getting me tripped up it seems.
match &* hand is necessary for it to recognize patterns in a Vector. There may be a better way of handling this.
To ensure that a hand_a or hand_b Vector can be sent to multiple functions, it seems cloning is necessary.